+++ src/adera/Machines/RCSController.h
@@ -47,28 +47,33 @@ public:
 
     static void add_functions(osp::active::ActiveScene& rScene);
 
+    /**
+     * Constructs MachineRCSControllers for vehicles in-construction
+     *
+     * @param rScene [ref] Scene supporting vehicles
+     */
     static void update_construct(osp::active::ActiveScene &rScene);
 
     /**
      * Primary system update function
-     * 
+     *
      * Iterates over MachineRCSControllers and issues throttle commands to the
      * associated MachineRocket thrusters via wire
-     * 
-     * @param rScene [in] - The scene to update
+     *
+     * @param rScene [ref] Scene with MachineRCSControllers to update
      */
     static void update_controls(osp::active::ActiveScene &rScene);
 
 private:
     /**
      * Command-thrust influence calculator
-     * 
+     *
      * Given a thruster's orientation and position relative to ship center of
      * mass, and a translation and rotation command, calculates how much
      * influence the thruster has on the commanded motion. Called on all
      * vehicle RCS thrusters to decide which are necessary to respond to the
      * maneuver command.
-     * 
+     *
      * @param posOset   [in] The position of the thruster relative to the ship CoM
      * @param direction [in] Direction that the thruster points
      * @param cmdTransl [in] Commanded translation vector
@@ -79,6 +84,8 @@ private:
         Magnum::Vector3 cmdTransl, Magnum::Vector3 cmdRot);
 
 }; // SysMachineRCSController
+
+//-----------------------------------------------------------------------------
 
 class MachineRCSController : public osp::active::Machine
 {
@@ -94,7 +101,7 @@ public:
     ~MachineRCSController() = default;
 
     void propagate_output(osp::active::WireOutput *output) override;
-    
+
     osp::active::WireInput* request_input(osp::WireInPort port) override;
     osp::active::WireOutput* request_output(osp::WireOutPort port) override;
