+++ src/adera/SysExhaustPlume.h
@@ -28,20 +28,20 @@
 #include "adera/Shaders/PlumeShader.h"
 #include "osp/Resource/Resource.h"
 
-namespace osp::active
+namespace adera::active
 {
 
 struct ACompExhaustPlume
 {
-    ActiveEnt m_parentMachineRocket{entt::null};
-    DependRes<PlumeEffectData> m_effect;
+    osp::active::ActiveEnt m_parentMachineRocket{entt::null};
+    osp::DependRes<PlumeEffectData> m_effect;
 };
 
 class SysExhaustPlume
 {
 public:
 
-    static void add_functions(ActiveScene& rScene);
+    static void add_functions(osp::active::ActiveScene& rScene);
 
     /**
      * Initialize plume graphics
@@ -50,9 +50,10 @@ public:
      * this function takes such entities, retrieves the appropriate graphics
      * resources, and configures the graphical components
      */
-    static void initialize_plume(ActiveScene& rScene, ActiveEnt e);
+    static void initialize_plume(osp::active::ActiveScene& rScene,
+                                 osp::active::ActiveEnt e);
 
-    static void update_plumes(ActiveScene& rScene);
+    static void update_plumes(osp::active::ActiveScene& rScene);
 };
 
-} // namespace osp::active
+} // namespace adera::active
